requirements.txt
bw2calc
bw2data
lxml
numpy
progressbar-ipython
scipy
stats_arrays
unicodecsv
unidecode
voluptuous
xlsxwriter
